adding article feedback component
