+++ src/stories/article-feedback.component.stories.ts
@@ -1,0 +1,14 @@
+import { Meta, StoryFn } from '@storybook/angular';
+import { ArticleFeedbackComponent } from '../app/components/shared/article-feedback/article-feedback.component';
+
+export default {
+  title: 'Shared/Article-Feedback',
+  component: ArticleFeedbackComponent,
+} as Meta;
+
+const Template: StoryFn = (args) => ({
+  props: args,
+});
+
+export const Default = Template.bind({});
+Default.args = {};
