+++ src/styles/mixins.scss
@@ -1,0 +1,27 @@
+@mixin respond-to($media) {
+    @if $media == "mobile" {
+        @media (max-width: 480px) {
+        @content;
+        }
+    } @else if $media == "tablet" {
+        @media (max-width: 768px) {
+        @content;
+        }
+    } @else if $media == "desktop" {
+        @media (min-width: 769px) {
+        @content;
+        }
+    }
+}
+
+@mixin respond-to($media) {
+    @if $media == "mobile-mb" {
+        @media (max-width: 480px) {
+        @content;
+        }
+    } @else if $media == "desktop-bg" {
+        @media (min-width: 481px) {
+        @content;
+        }
+    }
+}
